python-dashboard/requirements.txt
flask
flask-sqlalchemy
flask_jwt_extended
psycopg2
flask_migrate
flask-cors